BF: Drop --contain from the singularity run commands
The `singularity run` commands (with singularity version 2.3.2) throw
a "No such file or directory" error because --pwd has no effect when
it's passed with --contain (at least as of 2.3.2).  The --contain and
--pwd issue has been reported at singularityware/singularity#1208 .
In the meantime, drop --contain from the instructions.

Also, note that the current instructions fail with singularity 2.4
because the import command no longer exists.